# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smart WMS API",
        "description": "Warehouse Management System API for Android picking terminals",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://smart-wms.test",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login to WMS",
                "description": "Authenticate picker with code and password, returns API token",
                "operationId": "dd51a484b7fa0f97c7680fdbde14cf5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "code",
                                    "password"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "TEST001"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "device_id": {
                                        "type": "string",
                                        "example": "ANDROID-12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "LOGIN_SUCCESS"
                                        },
                                        "result": {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "token": {
                                                            "type": "string",
                                                            "example": "1|abcdef123456..."
                                                        },
                                                        "picker": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 2
                                                                },
                                                                "code": {
                                                                    "type": "string",
                                                                    "example": "TEST001"
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "テストピッカー"
                                                                },
                                                                "default_warehouse_id": {
                                                                    "type": "integer",
                                                                    "example": 991
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Login successful"
                                                },
                                                "debug_message": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials or inactive account",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "UNAUTHORIZED"
                                        },
                                        "result": {
                                            "properties": {
                                                "data": {
                                                    "type": "object",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "error_message": {
                                                    "type": "string",
                                                    "example": "Invalid credentials"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "VALIDATION_ERROR"
                                        },
                                        "result": {
                                            "properties": {
                                                "data": {
                                                    "type": "object",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "error_message": {
                                                    "type": "string",
                                                    "example": "Validation failed"
                                                },
                                                "errors": {
                                                    "properties": {
                                                        "code": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string",
                                                                "example": "validation.required"
                                                            }
                                                        },
                                                        "password": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string",
                                                                "example": "validation.required"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout from WMS",
                "description": "Revoke current API token",
                "operationId": "df9bdc6c7a6d9f20b983526b2ec0f698",
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "LOGOUT_SUCCESS"
                                        },
                                        "result": {
                                            "properties": {
                                                "data": {
                                                    "type": "object",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Logged out successfully"
                                                },
                                                "debug_message": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "UNAUTHORIZED"
                                        },
                                        "result": {
                                            "properties": {
                                                "data": {
                                                    "type": "object",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "error_message": {
                                                    "type": "string",
                                                    "example": "Unauthenticated"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": [],
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get current picker info",
                "description": "Returns information about the authenticated picker",
                "operationId": "8367feff40ba10e43187792cad3132c1",
                "responses": {
                    "200": {
                        "description": "Picker information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "SUCCESS"
                                        },
                                        "result": {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "code": {
                                                            "type": "string",
                                                            "example": "TEST001"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "テストピッカー"
                                                        },
                                                        "default_warehouse_id": {
                                                            "type": "integer",
                                                            "example": 991
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "debug_message": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "UNAUTHORIZED"
                                        },
                                        "result": {
                                            "properties": {
                                                "data": {
                                                    "type": "object",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "error_message": {
                                                    "type": "string",
                                                    "example": "Unauthenticated"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": [],
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "API Key for WMS API access (required for all endpoints)",
                "name": "X-API-Key",
                "in": "header"
            },
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "token",
                "description": "Bearer token from /auth/login endpoint"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        }
    ]
}